Fixed a few things, added status bar + file path indicator, added icons and more

+++ MCNBTEditor.Core/Explorer/NBT/TagDataFileViewModel.cs
@@ -10,27 +10,39 @@ using MCNBTEditor.Core.Views.Dialogs.Message;
 namespace MCNBTEditor.Core.Explorer.NBT {
     public class TagDataFileViewModel : TagCompoundViewModel, IHaveFilePath {
         private bool hasFileSavedOnce;
-
         private string filePath;
+        private bool isCompressed = true;
+        private bool isBigEndian = true;
+        private bool isModified;
+
+        /// <summary>
+        /// The file path of this tag
+        /// </summary>
         public string FilePath {
             get => this.filePath;
             set => this.RaisePropertyChanged(ref this.filePath, value);
         }
 
-        private bool isCompressed = true;
         public bool IsCompressed {
             get => this.isCompressed;
             set => this.RaisePropertyChanged(ref this.isCompressed, value);
         }
 
-        private bool isBigEndian = true;
         public bool IsBigEndian {
             get => this.isBigEndian;
             set => this.RaisePropertyChanged(ref this.isBigEndian, value);
         }
 
+        /// <summary>
+        /// Whether or not this tag has dirty data that the user may want to save
+        /// </summary>
+        public bool IsModified {
+            get => this.isModified;
+            set => this.RaisePropertyChanged(ref this.isModified, value);
+        }
+
         public AsyncRelayCommand RefreshDataCommand { get; }
-        public AsyncRelayCommand ShowInExplorerCommand { get; }
+        public AsyncRelayCommand OpenInExplorerCommand { get; }
         public AsyncRelayCommand DeleteFileCommand { get; }
         public RelayCommand CopyFilePathToClipboardCommand { get; }
         public AsyncRelayCommand SaveFileCommand { get; }
@@ -48,7 +60,7 @@ namespace MCNBTEditor.Core.Explorer.NBT {
         public TagDataFileViewModel(string name) : base(name) {
             this.RefreshDataCommand = new AsyncRelayCommand(this.RefreshAction, () => File.Exists(this.FilePath));
             this.DeleteFileCommand = new AsyncRelayCommand(this.DeleteFileAction, () => File.Exists(this.FilePath));
-            this.ShowInExplorerCommand = new AsyncRelayCommand(this.OpenInExplorerAction, () => IoC.ExplorerService != null && File.Exists(this.FilePath));
+            this.OpenInExplorerCommand = new AsyncRelayCommand(this.OpenInExplorerAction, () => IoC.ExplorerService != null && File.Exists(this.FilePath));
             this.SaveFileCommand = new AsyncRelayCommand(() => this.SaveToFileAction(false));
             this.SaveFileAsCommand = new AsyncRelayCommand(() => this.SaveToFileAction(true));
             this.CopyFilePathToClipboardCommand = new RelayCommand(() => {
@@ -82,20 +94,62 @@ namespace MCNBTEditor.Core.Explorer.NBT {
         }
 
         public async Task DeleteFileAction() {
-            bool canRemove = false;
+            string result = await Dialogs.RemoveItemWhenDeletingDialog.ShowAsync("Remove file from tree?", "Do you want to also remove the DAT file from the tree/list?");
+            if (result == "cancel") {
+                return;
+            }
+
             if (File.Exists(this.FilePath)) {
                 try {
                     File.Delete(this.FilePath);
-                    canRemove = true;
                 }
                 catch (Exception e) {
-                    await IoC.MessageDialogs.ShowMessageAsync("Failed to delete file", $"Failed to delete {this.filePath}:\n{e.Message}");
+                    result = "no";
+                    await IoC.MessageDialogs.ShowMessageExAsync("Failed to delete file", $"Failed to delete {this.filePath}", e.ToString());
                 }
             }
 
-            if (canRemove && "yes" == await Dialogs.RemoveDatFileWhenDeletingDialog.ShowAsync("Remove dat file?", "Do you want to also remove the DAT file from the list?")) {
-                await this.RemoveFromParentAsync(); // removes from root
+            if (result == "yes") {
+                this.RemoveFromParent(); // removes from root
             }
+        }
+
+        public override async Task RemoveFromParentAction() {
+            if (this.IsModified) {
+                string yesNoDialog = await Dialogs.YesNoCancelDialog.ShowAsync("Modified tag", "This .DAT file has been modified since you opened it. " + (string.IsNullOrEmpty(this.FilePath) ? "Do you want to save the changes?" : $"Do you want to save the changes to {this.FilePath}"));
+                if (yesNoDialog == "cancel") {
+                    return;
+                }
+                else if (yesNoDialog == "yes") {
+                    string path;
+                    if (string.IsNullOrEmpty(this.FilePath) || (!this.hasFileSavedOnce && !File.Exists(this.FilePath))) {
+                        DialogResult<string> result = IoC.FilePicker.ShowSaveFileDialog(Filters.NBTCommonTypesWithAllFiles, titleBar: "Select a save location for the NBT file");
+                        if (!result.IsSuccess || string.IsNullOrEmpty(result.Value)) {
+                            return;
+                        }
+
+                        path = result.Value;
+                    }
+                    else {
+                        path = this.FilePath;
+                    }
+
+                    try {
+                        this.SaveToFile(path);
+                        this.hasFileSavedOnce = true;
+                        if (!path.Equals(this.FilePath)) {
+                            this.FilePath = path;
+                        }
+
+                        this.Name = Path.GetFileName(path);
+                    }
+                    catch (Exception e) {
+                        await IoC.MessageDialogs.ShowMessageAsync("Failed to write NBT", $"Failed to write compressed NBT to file at:\n{this.FilePath}\n{e.Message}");
+                    }
+                }
+            }
+
+            await base.RemoveFromParentAction();
         }
 
         public Task OpenInExplorerAction() {
@@ -114,9 +168,9 @@ namespace MCNBTEditor.Core.Explorer.NBT {
             CompressedStreamTools.Write(this.ToNBT(), filePath, this.IsCompressed, this.IsBigEndian);
         }
 
-        public async Task SaveToFileAction(bool saveAs) {
+        public async Task SaveToFileAction(bool forceSaveAs) {
             string path;
-            if (saveAs || string.IsNullOrEmpty(this.FilePath) || (!this.hasFileSavedOnce && !File.Exists(this.FilePath))) {
+            if (forceSaveAs || string.IsNullOrEmpty(this.FilePath) || (!this.hasFileSavedOnce && !File.Exists(this.FilePath))) {
                 DialogResult<string> result = IoC.FilePicker.ShowSaveFileDialog(Filters.NBTCommonTypesWithAllFiles, titleBar: "Select a save location for the NBT file");
                 if (!result.IsSuccess || string.IsNullOrEmpty(result.Value)) {
                     return;
@@ -147,6 +201,14 @@ namespace MCNBTEditor.Core.Explorer.NBT {
             }
             catch (Exception e) {
                 await IoC.MessageDialogs.ShowMessageAsync("Failed to write NBT", $"Failed to write compressed NBT to file at:\n{this.FilePath}\n{e.Message}");
+            }
+        }
+
+        public void DoSaveFile(string path) {
+            this.SaveToFile(path);
+            this.hasFileSavedOnce = true;
+            if (!path.Equals(this.FilePath)) {
+                this.FilePath = path;
             }
         }
